Remove sysops as a dep
Sysops was used to get a list of servers to tail.  However, now that we
have a VPN (and internal VPN dns) this isn't necessary.  We can instead
grab the list of servers and infer the type from the name.  This is
faster and doesn't require Bubbles to have an IAM account.